Changes including doubles

+++ app/src/main/java/com/example/aabid/gittogether/data/User.kt
@@ -2,7 +2,7 @@ package com.example.aabid.gittogether.data
 
 data class User(var uid: String = "",
                 var name: String = "",
-                var latitude: Int = 0,
-                var longitude: Int = 0,
+                var latitude: Double = 0.0,
+                var longitude: Double = 0.0,
                 var groups: MutableList<String> = mutableListOf()
 )
